btrfs-progs: headers cleanup
There are quite some headers included but not utilized, clangd would
report those headers for cleanup.

However clangd (17.0.6 at the time of writing) has a bug that, the file
is only utilizing a C macro in the header, clangd would still report the
header as unused.

This involves the following exceptions:

- pretty_size()
  There is not much need to declare it as a macro.
  Redefine it as a static inline function to workaround the clangd bug.

- kernel-lib/mktables.c
  False alert on inittype.h, as some types are defined as macros.

- libbtrfs/send-utils.c
  False alert on limits.h, as some types are defined as macros.

- tests/array-tests.c
- tests/string-table-test.c
  False alert on common/utils.h for ARRAY_SIZE() macro.

Some parts are not touched yet:

- crypto/*
  This involves quite some optimization code for different instructions,
  making it pretty hard to comprehensively tested.

- tests/library-tests
  As I'll do a more comprehensive cleanup involving BTRFS_FLAT_INCLUDES.

Signed-off-by: Qu Wenruo <wqu@suse.com>