btrfs-progs: docs: Clarify btrfs-send checksum
The way the CRC32C checksum used for btrfs-send differs from the way it's used elsewhere in btrfs. Without making the distinction, it's easy to make the flawed assumption that CRC32C always refers to the same, and end up with code that produces the wrong checksums.

This small note should guide the reader to the right function.